+++ crates/lo_odf/src/lib.rs
@@ -9,7 +9,7 @@ mod text;
 pub use common::{ExtraFile, MIME_ODB, MIME_ODF, MIME_ODG, MIME_ODP, MIME_ODS, MIME_ODT};
 pub use database::save_database_document;
 pub use draw::save_drawing_document;
-pub use formula::save_formula_document;
+pub use formula::{save_formula_document, save_formula_document_bytes};
 pub use presentation::save_presentation_document;
 pub use spreadsheet::save_spreadsheet_document;
 pub use text::save_text_document;